Obtener transacciones desde App.jsx
