+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Login from './Login';
 import Dashboard from './Dashboard';
@@ -10,16 +10,78 @@ import Budget from './Budget';
 import Alerts from './Alerts';
 import { Link } from 'react-router-dom';
 // import handle from 'mqtt/lib/handlers/index';
+import PublicGoogleSheetsParser from 'public-google-sheets-parser';
+import axios from 'axios';
+import API_URL from './api';
 
 
 function App() {
   const [user, setUser] = useState(null);
   const [isOpen, setIsOpen] = useState(false);
-  
+  const [excel, setExcel] = useState([]);
+
+  // cargar el excel al cargar la página
+  useEffect(() => {
+    const loadExcel = async () => {
+      try {
+        // TODO: mover el id del sheet a una variable de entorno
+        const parser = new PublicGoogleSheetsParser('1LHrW5rWxMevVtGhdjNU56j_tk_wCzszYP5w2Beg-q0s')
+        const data = await parser.parse();
+        setExcel(data);
+        console.log(data);
+      } catch (error) {
+        console.log(error);
+      }
+
+    };
+
+    loadExcel();
+
+  }, []);
+
+  // elejir entrada transaccion random
+  // TODO: hacerlo secuencial con un contador_id
+  const chooseRandomEntry = () => {
+    if (excel.length > 0) {
+      const randomIndex = Math.floor(Math.random() * excel.length);
+      const entry = excel[randomIndex];
+      console.log('Entrada seleccionada:', entry);
+      createTransaction(entry);
+    } else {
+      alert("El excel no contiene datos o no ha sido cargado");
+    }
+  };
+
+  const createTransaction = async (entry) => {
+
+    try {
+      const response = await axios.post(`${API_URL}/transactions/${user.id}`,
+        {
+          entry,
+        },
+      );
+
+      if (response.status === 201) {
+        console.log("Transaccion creada exitosamente");
+        const newTransaction = response.data.newTransaction;
+        // setTransac((prev) => [...prev, newTransaction]);
+
+        // TODO: procesar la transaccion
+        //  revisar si se sobrepaso el presupuesto
+        //  enviar alertas
+      }
+    } catch (error) {
+      // alert("Error al crear la transaccion");
+      console.log(error);
+    }
+  };
+
   return (
     <Router>
       <header>
-        <h2>Nombre App</h2>
+        {/* Adornar el titulo una vez hayamos elejido un nombre */}
+        <h2><Link to='/home'>Nombre App</Link></h2>
+        <button className="button" onClick={chooseRandomEntry}>Elegir Entrada Aleatoria</button>
         {user && (
           <div>
             <button className="hamburger" onClick={() => setIsOpen(!isOpen)}>
@@ -37,7 +99,7 @@ function App() {
                 <Link to="/" onClick={() => setUser(null)}>Cerrar sesión</Link>
               </li>
             </ul>
-            </div>
+          </div>
         )}
 
       </header>
